Fix old transaction details json
