--- sdk/src/main/java/com/ingresse/sdk/model/response/BasketSessionsJSON.kt
@@ -1,5 +1,0 @@
-package com.ingresse.sdk.model.response
-
-data class BasketSessionsJSON(
-    val id: Int? = 0,
-    val dateTime: SessionDateJSON?)
