--- sdk/src/main/java/com/ingresse/sdk/model/response/PaymentCardJSON.kt
@@ -1,5 +1,0 @@
-package com.ingresse.sdk.model.response
-
-data class PaymentCardJSON(
-    val firstDigits: String? = "",
-    val lastDigits: String? = "")
