--- sdk/src/main/java/com/ingresse/sdk/model/response/RefundJSON.kt
@@ -1,6 +1,0 @@
-package com.ingresse.sdk.model.response
-
-data class RefundJSON(
-    val operator: String? = "",
-    val reason: String? = "",
-    val date: String? = "")
