--- sdk/src/main/java/com/ingresse/sdk/model/response/TransactionBasketJSON.kt
@@ -1,4 +1,0 @@
-package com.ingresse.sdk.model.response
-
-data class TransactionBasketJSON(
-    val tickets: Array<TransactionTicketJSON>?)
